+++ s4e-web/src/app/state/session/session.model.ts
@@ -1,7 +1,6 @@
 export interface LoginFormState {
-  login: string;
+  email: string;
   password: string;
-  rememberMe: boolean;
 }
 
 export interface Role {