+++ s4e-web/src/app/utils.ts
@@ -1,4 +1,4 @@
-import {UrlSegment} from '@angular/router';
+import { UrlSegment } from '@angular/router';
 
 export function activateMatcher(url: UrlSegment[]) {
   const uuidv4 = new RegExp(/^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/i);