+++ s4e-web/src/app/views/settings/people/invitation-form/invitation-form-modal.model.ts
@@ -1,7 +1,13 @@
+import { Institution } from 'src/app/views/settings/state/institution/institution.model';
 import {Modal} from '../../../../modal/state/modal.model';
 
 export const INVITATION_FORM_MODAL_ID = 'person-form-modal';
 
+export interface InvitationFormModal extends Modal {
+  institution: Institution;
+}
+
 export function isInvitationFormModal(modal: Modal): modal is Modal {
-  return modal.id === INVITATION_FORM_MODAL_ID;
+  return modal.id === INVITATION_FORM_MODAL_ID
+    && (modal as InvitationFormModal).institution != null;
 }